+++ app/src/main/java/com/cajetan/youtubeplus/SeekDialog.java
@@ -1,0 +1,21 @@
+package com.cajetan.youtubeplus;
+
+import android.app.Dialog;
+import android.app.DialogFragment;
+import android.app.TimePickerDialog;
+import android.os.Bundle;
+import android.view.LayoutInflater;
+
+public class SeekDialog extends DialogFragment {
+
+    @Override
+    public Dialog onCreateDialog(Bundle savedInstanceState) {
+        TimePickerDialog.Builder builder = new TimePickerDialog.Builder(getActivity());
+
+        LayoutInflater inflater = getActivity().getLayoutInflater();
+        builder.setView(inflater.inflate(R.layout.dialog_seek, null));
+
+        return builder.create();
+    }
+
+}